requirements.txt
python-dotenv-1.0.0
requests
aiohttp
asyncpg-0.28.0
SQLAlchemy-1.4.48
more_itertools-9.1.0